requirements.txt
streamlit
Pillow
numpy
torch
torchvision
transformers
sentence-transformers
faiss-cpu
opencv-python
